refactor: extract Console image controller
