+++ Tests/UI/test_console_native_chat_flow.py
@@ -23,6 +23,7 @@ from textual.widgets import Button, Checkbox, Input, Static, TextArea
 
 from Tests.fixtures.required_doubles import exploding_double
 from Tests.UI.background_signals import wait_for_background_signal, wait_for_signal
+from Tests.UI.console_controller_stubs import stub_image_controller
 from Tests.UI.test_destination_shells import _build_test_app, _wait_for_selector
 from Tests.UI.test_product_maturity_gate1_core_loop_screen_adaptation import (
     ConsoleHarness,
@@ -210,9 +211,7 @@ def test_console_workspace_conversation_search_worker_uses_dedicated_group():
     # contract this test pins (the search runs in its own exclusive worker
     # group, so a newer search cancels an in-flight one) is unchanged; only
     # the function that expresses it moved.
-    source = inspect.getsource(
-        ChatScreen._start_console_conversation_browser_search
-    )
+    source = inspect.getsource(ChatScreen._start_console_conversation_browser_search)
 
     assert 'group="console-workspace-conversation-search"' in source
     assert "exclusive=True" in source
@@ -1833,7 +1832,9 @@ def _console_conversation_browser_rows(console):
     always the dataclass default (``""``). Rebuilding the full context state
     is the seam that actually feeds the rendered row label.
     """
-    browser = console._workspace._build_console_workspace_context_state().conversation_browser
+    browser = (
+        console._workspace._build_console_workspace_context_state().conversation_browser
+    )
     rows = []
     for section in browser.sections:
         rows.extend(section.rows)
@@ -2869,9 +2870,7 @@ async def test_console_duplicate_send_during_stream_does_not_break_stop_control(
         assert send_button.label.plain == "Queue"
         controller = console._ensure_console_chat_controller()
         session_id = controller.store.active_session_id
-        assert (
-            controller.prompt_queue_registry.snapshot(session_id).total_count == 0
-        )
+        assert controller.prompt_queue_registry.snapshot(session_id).total_count == 0
 
         await console.handle_console_send_message(Button.Pressed(send_button))
         await pilot.pause(0.1)
@@ -7390,9 +7389,9 @@ async def test_console_conversation_browser_search_ignores_stale_results():
         # found. Rows: [stale-alpha]" while the state already held only
         # fresh-beta. Now a real regression fails here and names itself, and a
         # render that is merely late fails below saying so.
-        assert [row.conversation_id for row in console._console_conversation_browser_rows] == [
-            "fresh-beta"
-        ], "the stale search result overwrote the fresh one"
+        assert [
+            row.conversation_id for row in console._console_conversation_browser_rows
+        ] == ["fresh-beta"], "the stale search result overwrote the fresh one"
 
         await _wait_for_browser_conversation_row(console, pilot, "fresh-beta")
         row_texts = _console_workspace_conversation_texts(console)
@@ -8357,8 +8356,10 @@ async def test_console_workspace_conversation_search_shows_local_rows_before_slo
             await _wait_for_browser_render(
                 pilot,
                 lambda: "1 match" in _visible_text(console),
-                lambda: "'1 match' never rendered while the persisted search "
-                f"was still pending: {_visible_text(console)[:400]!r}",
+                lambda: (
+                    "'1 match' never rendered while the persisted search "
+                    f"was still pending: {_visible_text(console)[:400]!r}"
+                ),
             )
             assert not app.chat_conversation_scope_service.release.is_set()
             assert "1 match" in _visible_text(console)
@@ -8886,7 +8887,9 @@ async def test_console_resume_restores_server_character_identity_without_local_l
         console._resolve_resumed_character_name = local_lookup
 
         assert (
-            await console._workspace._resume_console_workspace_conversation("server-scoped")
+            await console._workspace._resume_console_workspace_conversation(
+                "server-scoped"
+            )
             is True
         )
         scoped = store.switch_session(store.active_session_id)
@@ -8900,7 +8903,9 @@ async def test_console_resume_restores_server_character_identity_without_local_l
         assert scoped.settings.character_label == ""
 
         assert (
-            await console._workspace._resume_console_workspace_conversation("server-unscoped")
+            await console._workspace._resume_console_workspace_conversation(
+                "server-unscoped"
+            )
             is True
         )
         unscoped = store.switch_session(store.active_session_id)
@@ -8986,7 +8991,9 @@ async def test_console_resume_rejects_character_identity_without_valid_source(
         console._resolve_resumed_character_name = local_lookup
 
         assert (
-            await console._workspace._resume_console_workspace_conversation("invalid-source")
+            await console._workspace._resume_console_workspace_conversation(
+                "invalid-source"
+            )
             is True
         )
 
@@ -9031,7 +9038,9 @@ async def test_console_resume_rehydrates_local_character_name_from_local_project
         console._resolve_resumed_character_name = name_lookup
 
         assert (
-            await console._workspace._resume_console_workspace_conversation("local-character")
+            await console._workspace._resume_console_workspace_conversation(
+                "local-character"
+            )
             is True
         )
 
@@ -9893,6 +9902,44 @@ def _bare_console_screen(store: ConsoleChatStore) -> ChatScreen:
     screen.app_instance = SimpleNamespace(
         notify=lambda *a, **k: None, chachanotes_db=None
     )
+    stub_image_controller(
+        screen,
+        context="test_console_native_chat_flow._bare_console_screen",
+        ensure_console_image_view=lambda: screen._ensure_console_image_view(),
+        recent_console_image_messages=(
+            lambda messages: screen._recent_console_image_messages(messages)
+        ),
+        console_image_default_mode=lambda: screen._console_image_default_mode,
+        console_generation_browse=lambda: screen._console_generation_browse(),
+        sync_native_console_chat_ui=lambda: screen._sync_native_console_chat_ui(),
+        ensure_console_chat_store=lambda: screen._ensure_console_chat_store(),
+        build_console_provider_selection=(
+            lambda: screen._build_console_provider_selection()
+        ),
+        ensure_console_provider_gateway=(
+            lambda: screen._ensure_console_provider_gateway()
+        ),
+        console_image_preparing=(
+            lambda: getattr(screen, "_console_image_preparing", None)
+        ),
+        current_console_chat_store=lambda: screen._console_chat_store,
+        console_composer_or_none=lambda: screen._console_composer_or_none(),
+        console_visible_draft_session_id=(
+            lambda: screen._console_visible_draft_session_id
+        ),
+        append_native_console_system_message=(
+            lambda *args, **kwargs: screen._append_native_console_system_message(
+                *args, **kwargs
+            )
+        ),
+        request_console_control_bar_sync=(
+            lambda: screen._request_console_control_bar_sync()
+        ),
+        default_console_session_settings=(
+            lambda: screen._session._default_console_session_settings()
+        ),
+        clear_console_composer_draft=(lambda: screen._clear_console_composer_draft()),
+    )
 
     # `_restore_native_console_state`'s message-rehydration calls
     # (`_rehydrate_console_message_image`/`_attachments`/`_generation_
@@ -9934,20 +9981,22 @@ def _bare_console_screen(store: ConsoleChatStore) -> ChatScreen:
         start_console_transcript_sync_timer=lambda: None,
         clear_native_console_message_selection=lambda: None,
         regenerate_console_generation_variant=(
-            lambda message_id: screen._regenerate_console_generation_variant(
+            lambda message_id: screen._image._regenerate_console_generation_variant(
                 message_id
             )
         ),
         select_console_generation_variant=(
-            lambda message, direction: screen._select_console_generation_variant(
+            lambda message, direction: screen._image._select_console_generation_variant(
                 message, direction=direction
             )
         ),
         keep_console_generation_variant=(
-            lambda message: screen._keep_console_generation_variant(message)
+            lambda message: screen._image._keep_console_generation_variant(message)
         ),
         handle_console_toggle_image_view=(
-            lambda message_id: screen._handle_console_toggle_image_view(message_id)
+            lambda message_id: screen._image._handle_console_toggle_image_view(
+                message_id
+            )
         ),
         invalidate_console_persisted_rows_cache=lambda: None,
     )
@@ -11427,19 +11476,19 @@ async def test_image_message_gets_inline_row_after_prep_and_toggle_cycles():
         assert console.query(f"#console-image-{message.id}"), "image row never appeared"
 
         # Toggle: pixels -> graphics (widget swaps, still present)
-        console._handle_console_toggle_image_view(message.id)
+        console._image._handle_console_toggle_image_view(message.id)
         await console._sync_native_console_chat_ui()
         await pilot.pause()
         assert console.query(f"#console-image-{message.id}")
 
         # Toggle: graphics -> hidden (row disappears)
-        console._handle_console_toggle_image_view(message.id)
+        console._image._handle_console_toggle_image_view(message.id)
         await console._sync_native_console_chat_ui()
         await pilot.pause()
         assert not console.query(f"#console-image-{message.id}")
 
         # Toggle: hidden -> pixels (row returns)
-        console._handle_console_toggle_image_view(message.id)
+        console._image._handle_console_toggle_image_view(message.id)
         await console._sync_native_console_chat_ui()
         await pilot.pause()
         assert console.query(f"#console-image-{message.id}")
@@ -11521,7 +11570,7 @@ def test_console_image_prep_bounded_to_cache_capacity_avoids_churn():
     # (a) + (b): specs are bounded to cache capacity and are the most recent
     # image messages — older messages were never prepared, so they can never
     # appear here regardless of how many messages the session holds.
-    specs = screen._build_console_image_specs(messages)
+    specs = screen._image._build_console_image_specs(messages)
     assert len(specs) <= IMAGE_CACHE_MAX_ENTRIES
     assert set(specs) == set(most_recent_ids)
 
@@ -12058,9 +12107,9 @@ async def test_console_save_as_savers_confirm_at_success_severity():
     success_toasts = [m for m, severity in notifications if severity == "success"]
     assert "Saved message as Note." in success_toasts
     assert "Saved message as Media. It appears under Library ▸ Media." in success_toasts
-    assert any(
-        m.startswith("Saved message as Prompt '") for m in success_toasts
-    ), success_toasts
+    assert any(m.startswith("Saved message as Prompt '") for m in success_toasts), (
+        success_toasts
+    )
     assert (
         "Saved message as a Chatbook artifact. It appears under Artifacts."
         in success_toasts
@@ -12100,9 +12149,7 @@ async def test_console_retry_accepted_fires_success_toast():
         await _wait_for_selector(
             console, pilot, f"#console-message-action-retry-{failed.id}"
         )
-        success_before = [
-            m for m, severity in notifications if severity == "success"
-        ]
+        success_before = [m for m, severity in notifications if severity == "success"]
         await pilot.click(f"#console-message-action-retry-{failed.id}")
         await _wait_for_text(console, pilot, "recovered")
 
@@ -12178,7 +12225,9 @@ async def test_console_routine_send_fires_no_success_toast():
     app = _build_test_app()
     app.chat_api_provider_value = "llama_cpp"
     app.chat_api_model_value = "test-model"
-    app.console_provider_gateway_factory = lambda: CapturingGateway(chunks=("hel", "lo"))
+    app.console_provider_gateway_factory = lambda: CapturingGateway(
+        chunks=("hel", "lo")
+    )
     notifications = _capture_notify_severities(app)
     host = ConsoleHarness(app)
