+++ tldw_chatbook/UI/Console_Modules/wiring.py
@@ -55,6 +55,7 @@ from tldw_chatbook.Widgets.Console.console_auto_speak_consent import (
 from .agent import ConsoleAgentController
 from .dictation import ConsoleDictationController
 from .hands_free import ConsoleHandsFreeController
+from .image import ConsoleImageController
 from .message import ConsoleMessageController
 from .prompt_queue import (
     ConsolePromptQueueUIController,
@@ -76,19 +77,20 @@ def build_console_controllers(
     rag_source_types_accessor: Callable[[], tuple[str, ...]],
     rag_top_k_accessor: Callable[[], int],
 ) -> None:
-    """Construct the Console screen's eight controllers and attach them.
+    """Construct the Console screen's nine controllers and attach them.
 
-    Assigns, in this order, `screen._workspace`, `screen._session`,
-    `screen._dictation`, `screen._hands_free`, `screen._message`,
-    `screen._prompts` and `screen._agent`. The order is documentation, not a
-    constraint: every cross-controller dependency below is resolved at call
-    time (see the module docstring), so no controller reads a sibling that
-    does not exist yet.
+    Assigns, in this order, `screen._image`, `screen._workspace`,
+    `screen._session`, `screen._dictation`, `screen._hands_free`,
+    `screen._message`, `screen._prompts`, `screen._agent`, and
+    `screen._prompt_queue`. The order is documentation, not a constraint:
+    every cross-controller dependency below is resolved at call time (see the
+    module docstring), so no controller reads a sibling that does not exist
+    yet.
 
     `ChatScreen.__init__` calls this at exactly the point the first
     construction used to occupy. That position matters: the ~250 attribute
     assignments around it in `__init__` include names these lambdas read, and
-    none of the seven constructors reads anything off `screen` eagerly (each
+    none of the nine constructors reads anything off `screen` eagerly (each
     stores `screen` and its callables and nothing else), so the call needs to
     sit where it can see everything the pre-move constructions could.
 
@@ -100,6 +102,45 @@ def build_console_controllers(
     Returns:
         None. The controllers are reachable as attributes of `screen`.
     """
+    screen._image = ConsoleImageController(
+        screen,
+        app_instance=screen.app_instance,
+        ensure_console_image_view=lambda: screen._ensure_console_image_view(),
+        recent_console_image_messages=(
+            lambda messages: screen._recent_console_image_messages(messages)
+        ),
+        console_image_default_mode=lambda: screen._console_image_default_mode,
+        console_generation_browse=lambda: screen._console_generation_browse(),
+        sync_native_console_chat_ui=lambda: screen._sync_native_console_chat_ui(),
+        ensure_console_chat_store=lambda: screen._ensure_console_chat_store(),
+        build_console_provider_selection=(
+            lambda: screen._build_console_provider_selection()
+        ),
+        ensure_console_provider_gateway=(
+            lambda: screen._ensure_console_provider_gateway()
+        ),
+        console_image_preparing=(
+            lambda: getattr(screen, "_console_image_preparing", None)
+        ),
+        current_console_chat_store=lambda: screen._console_chat_store,
+        console_composer_or_none=lambda: screen._console_composer_or_none(),
+        console_visible_draft_session_id=(
+            lambda: screen._console_visible_draft_session_id
+        ),
+        append_native_console_system_message=(
+            lambda *args, **kwargs: screen._append_native_console_system_message(
+                *args, **kwargs
+            )
+        ),
+        request_console_control_bar_sync=(
+            lambda: screen._request_console_control_bar_sync()
+        ),
+        default_console_session_settings=(
+            lambda: screen._session._default_console_session_settings()
+        ),
+        clear_console_composer_draft=(lambda: screen._clear_console_composer_draft()),
+    )
+
     #: Workspace policy context, lifecycle, and resume-flow state and
     #: behaviour moved to `ConsoleWorkspaceController` (wave-2 console
     #: decomposition, task 2) -- the largest cluster in wave 2.
@@ -143,16 +184,12 @@ def build_console_controllers(
         default_session_settings_accessor=(
             lambda: screen._session._default_console_session_settings()
         ),
-        scope_picker_listers_accessor=(
-            lambda: screen._console_scope_picker_listers()
-        ),
+        scope_picker_listers_accessor=(lambda: screen._console_scope_picker_listers()),
         active_native_session_accessor=(
             lambda: screen._session._active_native_console_session()
         ),
         refresh_effective_scope_and_sync=(
-            lambda session: screen._refresh_console_effective_scope_and_sync(
-                session
-            )
+            lambda session: screen._refresh_console_effective_scope_and_sync(session)
         ),
         # Message <-> workspace seam (the reverse direction of the
         # session/message seams the message controller's own
@@ -164,9 +201,7 @@ def build_console_controllers(
         # resolves `self._message` at CALL time, so construction order
         # (workspace built before message) does not matter.
         messages_from_conversation_tree_accessor=(
-            lambda tree: screen._message._console_messages_from_conversation_tree(
-                tree
-            )
+            lambda tree: screen._message._console_messages_from_conversation_tree(tree)
         ),
         session_settings_for_resume_accessor=(
             lambda conversation: screen._session._console_session_settings_for_resume(
@@ -174,9 +209,7 @@ def build_console_controllers(
             )
         ),
         resolve_resumed_character_name=(
-            lambda character_id: screen._resolve_resumed_character_name(
-                character_id
-            )
+            lambda character_id: screen._resolve_resumed_character_name(character_id)
         ),
         # Agent <-> workspace seam, same shape as the message seam above:
         # the resume flow's TOOL-marker re-derivation moved to
@@ -185,17 +218,13 @@ def build_console_controllers(
         # changed, from the screen's own method to the controller.
         inject_resume_agent_markers_accessor=(
             lambda messages, conversation_id: (
-                screen._agent._inject_resume_agent_markers(
-                    messages, conversation_id
-                )
+                screen._agent._inject_resume_agent_markers(messages, conversation_id)
             )
         ),
         resolve_effective_scope_state=(
             lambda session: screen._resolve_console_effective_scope_state(session)
         ),
-        sync_retrieval_scope_row=(
-            lambda: screen._sync_console_retrieval_scope_row()
-        ),
+        sync_retrieval_scope_row=(lambda: screen._sync_console_retrieval_scope_row()),
         note_follow_intent=lambda: screen._note_console_follow_intent(),
         focus_composer_if_needed=(
             lambda **kwargs: screen._focus_console_composer_if_needed(**kwargs)
@@ -245,9 +274,7 @@ def build_console_controllers(
         effective_console_provider_model=(
             lambda: screen._effective_console_provider_model()
         ),
-        provider_readiness_app_config=(
-            lambda: screen._provider_readiness_app_config()
-        ),
+        provider_readiness_app_config=(lambda: screen._provider_readiness_app_config()),
         build_provider_selection=(
             lambda session_id: screen._build_console_provider_selection(session_id)
         ),
@@ -272,9 +299,7 @@ def build_console_controllers(
             )
         ),
         refresh_effective_scope_and_sync=(
-            lambda session: screen._refresh_console_effective_scope_and_sync(
-                session
-            )
+            lambda session: screen._refresh_console_effective_scope_and_sync(session)
         ),
         # Session<->workspace seam (design spec: "a named callable
         # between them; design it deliberately, never a back-door
@@ -283,9 +308,7 @@ def build_console_controllers(
         # so construction order does not matter.
         set_active_workspace_for_session=(
             lambda session_id: (
-                screen._workspace._set_active_workspace_for_console_session(
-                    session_id
-                )
+                screen._workspace._set_active_workspace_for_console_session(session_id)
             )
         ),
         resume_workspace_conversation=(
@@ -363,9 +386,7 @@ def build_console_controllers(
             )
         ),
         enter_hands_free_loop=(
-            lambda **kwargs: screen._hands_free._enter_console_hands_free_loop(
-                **kwargs
-            )
+            lambda **kwargs: screen._hands_free._enter_console_hands_free_loop(**kwargs)
         ),
         hands_free_force_immediate_send=(
             lambda: screen._hands_free._console_hands_free_force_immediate_send()
@@ -381,9 +402,7 @@ def build_console_controllers(
         # was the same "temporary exception" shape as the four above,
         # for the same staleness reason, now closed out the same way.
         realtime_adopt_transcript=(
-            lambda transcript: screen._console_realtime_adopt_transcript(
-                transcript
-            )
+            lambda transcript: screen._console_realtime_adopt_transcript(transcript)
         ),
         # Same screen-owned realtime engine, read as a live session rather
         # than called: `_handle_console_dictation_button` (wave-4 task 2)
@@ -392,9 +411,7 @@ def build_console_controllers(
         # Controller` takes an identically-named, identically-shaped one.
         realtime_session_accessor=lambda: screen._console_realtime,
         run_pending_voice_action=(
-            lambda session_id: screen._run_pending_console_voice_action(
-                session_id
-            )
+            lambda session_id: screen._run_pending_console_voice_action(session_id)
         ),
         undo_histories_accessor=lambda: screen._console_undo_histories,
         visible_draft_session_id_accessor=(
@@ -431,9 +448,7 @@ def build_console_controllers(
         request_dictation_start=lambda: screen._request_console_dictation_start(),
         request_dictation_stop=lambda: screen._request_console_dictation_stop(),
         run_pending_voice_action=(
-            lambda session_id: screen._run_pending_console_voice_action(
-                session_id
-            )
+            lambda session_id: screen._run_pending_console_voice_action(session_id)
         ),
         realtime_session_accessor=lambda: screen._console_realtime,
         enter_realtime_loop=(
@@ -466,9 +481,7 @@ def build_console_controllers(
         ensure_console_chat_controller=(
             lambda: screen._ensure_console_chat_controller()
         ),
-        current_chat_controller_accessor=(
-            lambda: screen._console_chat_controller
-        ),
+        current_chat_controller_accessor=(lambda: screen._console_chat_controller),
         sync_native_console_chat_ui=lambda: screen._sync_native_console_chat_ui(),
         # Session <-> message seam (design spec: "a named callable
         # between them; design it deliberately, never a back-door
@@ -522,27 +535,27 @@ def build_console_controllers(
             lambda: screen._clear_native_console_message_selection()
         ),
         regenerate_console_generation_variant=(
-            lambda message_id: screen._regenerate_console_generation_variant(
+            lambda message_id: screen._image._regenerate_console_generation_variant(
                 message_id
             )
         ),
         select_console_generation_variant=(
-            lambda message, direction: screen._select_console_generation_variant(
+            lambda message, direction: screen._image._select_console_generation_variant(
                 message, direction=direction
             )
         ),
         keep_console_generation_variant=(
-            lambda message: screen._keep_console_generation_variant(message)
+            lambda message: screen._image._keep_console_generation_variant(message)
         ),
         handle_console_toggle_image_view=(
-            lambda message_id: screen._handle_console_toggle_image_view(message_id)
+            lambda message_id: screen._image._handle_console_toggle_image_view(
+                message_id
+            )
         ),
         invalidate_console_persisted_rows_cache=(
             lambda: screen._invalidate_console_persisted_rows_cache()
         ),
-        play_console_video=(
-            lambda message_id: screen._play_console_video(message_id)
-        ),
+        play_console_video=(lambda message_id: screen._play_console_video(message_id)),
         save_console_video_copy=(
             lambda message_id: screen._save_console_video_copy(message_id)
         ),
@@ -630,8 +643,8 @@ def build_console_controllers(
             lambda: screen._session._ensure_active_console_session_settings()
         ),
         apply_console_session_system_prompt=(
-            lambda system_prompt: (
-                screen._session._apply_console_session_system_prompt(system_prompt)
+            lambda system_prompt: screen._session._apply_console_session_system_prompt(
+                system_prompt
             )
         ),
         sync_console_session_draft=(
@@ -646,9 +659,7 @@ def build_console_controllers(
         ensure_console_provider_gateway=(
             lambda: screen._ensure_console_provider_gateway()
         ),
-        console_provider_blocker_copy=(
-            lambda: screen._console_provider_blocker_copy()
-        ),
+        console_provider_blocker_copy=(lambda: screen._console_provider_blocker_copy()),
         # A bare-attribute READ, not a call: the modal opener hands
         # this straight to `ConsolePromptsModal(configure_provider=...)`
         # without calling it, so the accessor must return the screen's
@@ -659,9 +670,7 @@ def build_console_controllers(
         open_console_provider_recovery_accessor=(
             lambda: screen._open_console_provider_recovery
         ),
-        console_setup_blocked_reason=(
-            lambda: screen._console_setup_blocked_reason()
-        ),
+        console_setup_blocked_reason=(lambda: screen._console_setup_blocked_reason()),
         focus_console_composer_if_needed=(
             lambda **kwargs: screen._focus_console_composer_if_needed(**kwargs)
         ),
@@ -673,9 +682,7 @@ def build_console_controllers(
                 text, replace=replace
             )
         ),
-        clear_console_composer_draft=(
-            lambda: screen._clear_console_composer_draft()
-        ),
+        clear_console_composer_draft=(lambda: screen._clear_console_composer_draft()),
         append_native_console_system_message=(
             lambda text: screen._append_native_console_system_message(text)
         ),
@@ -706,9 +713,7 @@ def build_console_controllers(
         screen,
         app_instance=screen.app_instance,
         chat_store_accessor=lambda: screen._ensure_console_chat_store(),
-        provider_gateway_accessor=(
-            lambda: screen._ensure_console_provider_gateway()
-        ),
+        provider_gateway_accessor=(lambda: screen._ensure_console_provider_gateway()),
         # A bare-attribute READ, not a call: `ConsoleAgentBridge` stores
         # this callable and calls it per run, so the accessor must return
         # the screen's method rather than today's answer -- the same shape
@@ -737,31 +742,21 @@ def build_console_controllers(
         ),
     )
     screen._prompt_queue = ConsolePromptQueueUIController(
-        chat_controller_accessor=(
-            lambda: screen._ensure_console_chat_controller()
-        ),
+        chat_controller_accessor=(lambda: screen._ensure_console_chat_controller()),
         ensure_active_session=(
             lambda: screen._session._ensure_active_console_session_settings()
         ),
         blocked_reason_accessor=lambda: screen._console_send_blocked_reason(),
-        setup_blocked_reason_accessor=(
-            lambda: screen._console_setup_blocked_reason()
-        ),
+        setup_blocked_reason_accessor=(lambda: screen._console_setup_blocked_reason()),
         restore_stash=lambda stash: screen._restore_console_send_stash(stash),
         append_system_message=(
             lambda text: screen._append_native_console_system_message(text)
         ),
         notify=(
-            lambda text, severity: screen.app_instance.notify(
-                text, severity=severity
-            )
+            lambda text, severity: screen.app_instance.notify(text, severity=severity)
         ),
-        focus_composer=(
-            lambda: screen._focus_console_composer_if_needed(force=True)
-        ),
-        inflight_stashes_accessor=(
-            lambda: screen._console_inflight_send_stashes
-        ),
+        focus_composer=(lambda: screen._focus_console_composer_if_needed(force=True)),
+        inflight_stashes_accessor=(lambda: screen._console_inflight_send_stashes),
         note_follow_intent=lambda: screen._note_console_follow_intent(),
         launch_chain=(
             lambda draft, session_id: screen.run_worker(
